autoresearch-trading/pyproject.toml
[project]
name = "autoresearch-trading"
version = "0.1.0"
requires-python = ">=3.12"
description = "Autoresearch trading system with LLM-driven strategy optimization"
authors = [{name = "Autoresearch"}]
dependencies = [
    "MetaTrader5",
    "vectorbt",
    "pandas",
    "numpy",
    "ta-lib",
    "pandas-ta",
    "openai",
    "loguru",
    "sqlalchemy",
    "apscheduler",
    "python-dotenv",
    "psutil",
    "rich",
    "pyarrow>=23.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"